% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{do}
\alias{do}
\title{do}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 17 rows and 4 columns.
}
\source{
http://www.gisdeveloper.co.kr/?p=2332
}
\usage{
do
}
\description{
Administrative districts of Korea shapefile, sido
}
\keyword{datasets}
