% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emd}
\alias{emd}
\title{emd}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 5047 rows and 4 columns.
}
\source{
http://www.gisdeveloper.co.kr/?p=2332
}
\usage{
emd
}
\description{
Administrative districts of Korea shapefile, eupmyeondong
}
\keyword{datasets}
