% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sgg}
\alias{sgg}
\title{sgg}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 250 rows and 4 columns.
}
\usage{
sgg
}
\description{
Administrative districts of Korea shapefile, sigungu
}
\keyword{datasets}
